+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Bumped `@actions/core` from 1.2.2 to 1.2.7
+- Added a new `use-workspace` input to allow loading a different version of the problem matcher that works with `dotnet-format`'s workspace log output. The allowed values are `true` and `false` (default).
 
 ## Version 1.2.0